fix(tests): fix broken set() calls from bitcast migration and ASAN workflow path
Three root causes for CI failures on main:

1. Botched bitcast→set() migration (Pattern A): 13 files had inline
   comments inside function call parentheses, e.g.:
     x.set(0, Float32(-100.0  # comment))
   The '#' starts a line comment, consuming the closing '))' and causing
   'expected ) in call argument list' parse errors. Fixed by moving
   comments outside the closing parentheses.

2. Botched bitcast→set() migration (Pattern B/C): test_backward_conv_padding
   had Float32(()) with value on next line; test_rmsprop had garbled
   Float32(= expr); test_optimizers had declarations commented out but
   usage left uncommented; test_core_operations had Int32(Int32()) with
   value on next line.

3. ASAN workflow referenced non-existent path 'tests/shared/core/types'
   (directory does not exist). Fixed to use 'tests/shared/core' with the
   dtype-specific test file pattern.

Affected groups: Core Utilities A, Core Utilities D, Core Gradient,
Core Activations & Types, Shared Infra & Testing, Misc Tests (compile
errors); ASAN Tests (wrong path).

+++ tests/shared/core/test_backward_conv_padding.mojo
@@ -62,9 +62,7 @@ fn test_conv2d_backward_grad_input_padding1() raises:
     var output = forward_input(x)
     var grad_output = zeros_like(output)
     for i in range(16):
-        grad_output.set(i, Float32(())
-            Float32(i % 4) * Float32(0.25) - Float32(0.3)
-        )
+        grad_output.set(i, Float32(i % 4) * Float32(0.25) - Float32(0.3))
 
     check_gradient(
         forward_input, backward_input, x, grad_output, rtol=1e-2, atol=1e-2
@@ -116,9 +114,7 @@ fn test_conv2d_backward_grad_weights_padding1() raises:
     var output = forward_weights(kernel)
     var grad_output = zeros_like(output)
     for i in range(16):
-        grad_output.set(i, Float32(())
-            Float32(i % 4) * Float32(0.25) - Float32(0.3)
-        )
+        grad_output.set(i, Float32(i % 4) * Float32(0.25) - Float32(0.3))
 
     check_gradient(
         forward_weights,
@@ -176,9 +172,7 @@ fn test_conv2d_backward_grad_input_padding2() raises:
     var output = forward_input(x)
     var grad_output = zeros_like(output)
     for i in range(49):
-        grad_output.set(i, Float32(())
-            Float32(i % 4) * Float32(0.25) - Float32(0.3)
-        )
+        grad_output.set(i, Float32(i % 4) * Float32(0.25) - Float32(0.3))
 
     check_gradient(
         forward_input, backward_input, x, grad_output, rtol=1e-2, atol=1e-2
@@ -231,9 +225,7 @@ fn test_conv2d_backward_grad_weights_padding2() raises:
     var output = forward_weights(kernel)
     var grad_output = zeros_like(output)
     for i in range(49):
-        grad_output.set(i, Float32(())
-            Float32(i % 4) * Float32(0.25) - Float32(0.3)
-        )
+        grad_output.set(i, Float32(i % 4) * Float32(0.25) - Float32(0.3))
 
     check_gradient(
         forward_weights,
